+++ README.md
@@ -1,4 +1,9 @@
 # SuperGrid
 A super simple jQuery plugin to render a grid out of an array of data.
 ***************
-#This is not working software yet.
+
+To run:
+> npm install
+> gulp serve
+
+- Deployed to http://localhost:8080/index.html
