+++ app/index.html
@@ -3,15 +3,15 @@
     <head>
         <meta charset="utf-8">
         <meta http-equiv="x-ua-compatible" content="ie=edge">
-        <title></title>
+        <title>Supergrid</title>
         <meta name="description" content="">
         <meta name="viewport" content="width=device-width, initial-scale=1">
 
-        <link rel="apple-touch-icon" href="apple-touch-icon.png">
+        <link rel="apple-touch-icon" href="img/apple-touch-icon.png">
         <!-- Place favicon.ico in the root directory -->
 
         <link rel="stylesheet" href="css/normalize.css">
-        <link rel="stylesheet" href="css/simpleGrid.css">
+        <link rel="stylesheet" href="css/supergrid.css">
         <link rel="stylesheet" href="css/main.css">
         <script src="js/vendor/modernizr-2.8.3.min.js"></script>
     </head>
@@ -30,11 +30,9 @@
         <script>window.jQuery || document.write('<script src="js/vendor/jquery-1.12.0.min.js"><\/script>')</script>
         <script src="https://code.jquery.com/ui/1.12.0/jquery-ui.min.js" integrity="sha256-eGE6blurk5sHj+rmkfsGYeKyZx3M4bG+ZlFyA7Kns7E=" crossorigin="anonymous"></script>
         <script>$.widget || document.write('<script src="js/vendor/jquery-1.12.0-ui.min.js"><\/script>')</script>
-        <script src="js/plugins.js"></script>
 
-        <script src="js/plugins.js"></script>
         <!-- Grid Plugin -->
-        <script src="js/simpleGrid.js"></script>
+        <script src="js/supergrid/supergrid.js"></script>
         <!-- Plugin Usage File -->
         <script src="js/main.js"></script>
