--- css/simpleGrid.css
@@ -1,67 +1,0 @@
-.supergrid {
-    border-bottom-width: 1px;
-    border-bottom-style: solid;
-    border-bottom-color: #DDDDDD;
-}
-
-.supergrid td {
-    padding: 5px;
-    border-width: 1px 1px 1px 1px;
-    border-style: solid;
-    border-color: #DDDDDD;
-    vertical-align: top;
-    white-space: normal;
-}
-
-.supergrid td.center {
-    text-align: center;
-}
-
-.supergrid .supergrid_header td {
-    white-space: nowrap;
-    position: relative;
-}
-
-.supergrid .supergrid_header td[data-sort] {
-    background-color: white;
-}
-
-/*.supergrid .supergrid_header td[data-sortable="true"][data-sort] .sort-icon {
-    display: inline-block;
-    background-image: url(../../images/main-sprite.png);
-    background-size: 400px 440px;
-    background-position: 100px 320px;
-    width: 15px;
-    height: 15px;
-    position: absolute;
-    right: 5px;
-    top: 7px;
-}*/
-
-.supergrid .supergrid_header td[data-sortable="true"][data-sort="asc"] .sort-icon {
-    background-position: 100px 320px;
-}
-
-.supergrid .supergrid_header td[data-sortable="true"][data-sort="desc"] .sort-icon {
-    background-position: 180px 320px;
-}
-
-.supergrid thead tr > td {
-    background: #ececec;
-}
-
-.supergrid tbody tr:nth-child(even) {
-    background-color: rgba(87, 126, 193, .1);
-}
-
-.supergrid tbody tr:nth-child(odd) {
-    background-color: white;
-}
-
-.supergrid tbody tr.hidden {
-    display: none;
-}
-
-.supergrid tbody .section td {
-  overflow:hidden;
-}
