--- js/min/main-min.js
@@ -1,1 +1,0 @@
-function getSampleData(e){var r=$.Deferred();return $.ajax({type:"GET",dataType:"json",url:e,success:r.resolve,error:r.reject}),r.promise()}function errorHandle(){alert("An Error Occurred Loading Local JSON")}function renderGrid(e,r){$(".grid").SuperGrid({columns:e[0]||[],data:r[0]||[]})}function generateData(){function e(){var e=a(0,10),r=["Advisory","Simple Application","Mono-Chromo Plastic","Warning Return","Basic Return","Razor Accelerated Death","Caffinated Dogs","Glitchy Gas","Holy Cow Collection","Holy Cow Distribution","Pea Pod"];return r[e]}function r(e,r){return new Date(e.getTime()+Math.random()*(r.getTime()-e.getTime()))}function a(e,r){return Math.floor(Math.random()*(r-e+1)+e)}function t(){return i={},i.id=a(1e5,9e5).toString(),i.login="aaaa",i.timeStamp=r(new Date(2012,0,1),new Date).getTime(),i.aarpNumber=a(100,2e3),i.docCode=Math.random().toString(36).slice(2,6),i.codedSerial=a(1e3,9999),i.title=e(),i}for(var n=[],o=0,i;1e3>o;)n.push(t()),o++;return n}$.when(getSampleData("/js/sampleData/config.json"),getSampleData("/js/sampleData/data.json")).then(renderGrid,errorHandle);
